requirements.txt
pygame>=2.5.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.8.0
rembg>=2.0.0
